Split generic arity annotation goldens
